Stop building seahub
Despite my best efforts, I cannot package all of seahub's dependencies
in a proper packaged way, and we need to run the pip install on the
installation machine, and therefore need python-dev.

In this case we may as well just avoid creating a proper package and
just use a simple shell script.